generate: improve enum tests
We shouldn’t test the Last enum case as that makes us sensitive to testing on a different version of vips than the one we generate on. This happens if the brew version is newer than the Linux version. The Last enum is always last, and isn’t useful for anything else. So if a new enum value is added, Last will change.

Also ignore deprecated enums that don’t match for some reason.